requirements.txt
moviepy==1.0.3
opencv-python-headless==4.8.1.78
numpy<2.0
Pillow==9.4.0
